Introduce the TestingContainer class.
This is a simplified version of the old ExtendedContainer.
When unit tests run an instance of this class will replace the
instance of RuntimeContainer used internally by Container.

If the environment variable USE_OLD_DI_CONTAINER is set
then ExtendedContainer will be used for unit tests instead
(this mechanism will be removed in WooCommerce 10.0).