Store customization: add missing context for "Paid" string
In languages other than English, "Paid" is translated differently whether it's referring to something that has been paid, or something that is not free.

Let's provide the necessary context to avoid mistranslations. In this case, the label was translated to "has been paid" in the interface in French for example, because of the other "paid" string in the codebase.