wp-env: Make mappings exclusive to test container
In the .wp-env.json file, we have several items in the mappings property
some of which add plugins to the WordPress site. However, these items
are all specifically for e2e testing and are not necessary for
development. In this PR I'm just moving the mappings definitions so that
they are specific to the testing container, and no longer also appear
in the development container.