Update customer invoice emails for failed payments
This updates the email templates for the customer invoice email (aka
"Send order details" email) to have different wording at the top when
the order has a failed status. The previous wording could be confusing
since it asks for payment but doesn't explain that their previous
payment attempt failed.

Fixes #52883