PaymentInfo: Abstract the exception handler
In `get_wcpay_card_info` we use some code from the WooPayments
extension. That code can potentially throw an exception, but it might
not be a standard Exception. It might be an API_Exception. We don't
need to differentiate our catch code between exception types, so it's
better for us to just use the abstract Throwable type, which will
catch all exceptions.