Reduce scope of globalized order metabox functions
In #51185 we exported the order metabox scripts outside of their
encapsulation in order to let 3rd party scripts re-initialize UI
elements if they were modified. However, this also opened up many other
methods to the global scope that don't need to be global. This amends
that change so that only the init methods are available globally.