Revert #51185 (#54390)
There are some concerns about de-encapsulating the objects in the
meta-boxes-order.js script and making them globally available. In the PR
that made this change, the goal was to be able to re-initialize elements
in the UI that had been replaced and had their event listeners deleted.
There are other ways to accomplish this without exposing the script
objects, so we should go back to the drawing board and try something
different.