Ensure GTIN,UPC, EAN, or ISBN field never starts or ends with a dash
This is a fix to ensure that on product edit page, GTIN,UPC, EAN, or ISBN field does not allow for a value that starts or ends with a dash. It extends the existing functionality from replacing invalid characters with empty strings  to making sure dashes at the begining and end are stripped too.

This is necessary because the existing solution allows values such as ABC-123 or 123-ABC which are not valid inputs for GTIN,UPC, EAN, or ISBN.