fix: prevent order statuses only on other admin pages
Order status are queried from stat table on every WC admin page and passed to Analytics Settings page. When the number of orders are large, making this query when the its cache expires is slow. For this reason this change is made to prevent this query from running on every WooCommerce admin page and made only on the Analytics settings page when it is required.