Avoid redirection loop in product canonical urls
In #51637 we introduced `wc_product_canonical_redirect` to ensure that
the category part of a product URL couldn't be any arbitrary string.
However, in the string comparison that happens within that method, we
didn't take into account that category slugs with multibyte characters
(e.g. Cyrillic or Arabic) that we retrieve with
`wc_product_post_type_link` would be url-encoded, while the category
slug from the query object would not. Thus we ended up with a situation
where there could be an infinite redirect loop. This change simply
ensures that both sides of the string comparison are decoded beforehand.

Fixes #54737
Might fix #54188